Rename pkg/oauth to pkg/oauthproto and move DCR primitives
Implements Phase 1 of issue #4977:
- Rename pkg/oauth/ -> pkg/oauthproto/; update all 17 import sites
  (8 with alias can now drop the alias since pkg name matches)
- Move DCR types, ScopeList, RegisterClientDynamically, and helpers
  from pkg/auth/oauth/dynamic_registration.go into pkg/oauthproto/dcr.go
- RegisterClientDynamically now takes *http.Client directly (nil builds
  a default); drop registerClientDynamicallyWithClient and HTTPClient iface
- Relocate IsLocalhost logic into pkg/oauthproto/IsLoopbackHost; pkg/networking
  keeps its IsLocalhost as a thin wrapper to avoid touching all callers
- Move UserAgent constant from pkg/auth/oauth/oidc.go to pkg/oauthproto/constants.go
- Generalize handleHTTPResponse error string (remove CLI-flag hints)
- Host NewDynamicClientRegistrationRequest in pkg/auth/discovery/dcr_request.go
  (sole caller; loopback redirect-URI assumption stays out of protocol pkg)
- Add tests in pkg/oauthproto/dcr_test.go covering the moved behaviour
- Update pkg/oauthproto/doc.go to document the expanded surface area

Address code review feedback

Fixed issues from code review:
- CRITICAL: Delete orphaned pkg/oauth/ directory (7 files); no importers remain
- HIGH: Copy request in RegisterClientDynamically before validateAndSetDefaults
  to avoid mutating the caller's struct; nil guard moved up before the deref
- HIGH: Drop redundant oauthproto alias from 4 test files (token_test.go,
  oidc_test.go, discovery_test.go, upstream/oidc_test.go)
- MEDIUM: Remove isLoopbackHost private wrapper; dcr.go calls IsLoopbackHost directly
- MEDIUM: Replace 13-line Apache header in pkg/oauthproto/doc.go with 2-line SPDX
- MEDIUM: Add comment in registerDynamicClient clarifying CLI-flag hint is
  intentional; reference #4978 for future structured error handling

Address second round of code review feedback

Fixed issues from code review:
- HIGH: Replace 13-line Apache header and redundant package doc comment in
  pkg/oauthproto/constants.go with the 2-line SPDX form; the file was
  substantially rewritten (UserAgent added, original URN constants removed)
  so it qualifies as "rewritten in full" per the project header rule
- MEDIUM: Drain response body with io.Copy(io.Discard) before returning the
  content-type error in handleHTTPResponse, allowing net/http to reuse the
  underlying TCP connection per the project drain-before-close style rule

Add missing io.LimitReader body-cap test for handleHTTPResponse

The Testing Strategy in issue #4977 requires that handleHTTPResponse is
tested for its 1 MB io.LimitReader cap. The non-JSON content-type and
missing client_id cases were already covered; this commit adds the cap
test. The test sends a body padded well past 1 MB and verifies the call
returns promptly without hanging or panicking — a future refactor that
removes or raises the cap would break the JSON decode and surface the
regression immediately.
