+++ pkg/auth/tokenexchange/middleware.go
@@ -15,7 +15,7 @@ import (
 	"golang.org/x/oauth2"
 
 	"github.com/stacklok/toolhive/pkg/auth"
-	"github.com/stacklok/toolhive/pkg/oauth"
+	"github.com/stacklok/toolhive/pkg/oauthproto"
 	"github.com/stacklok/toolhive/pkg/transport/types"
 )
 
@@ -72,8 +72,8 @@ type Config struct {
 	Scopes []string `json:"scopes,omitempty"`
 
 	// SubjectTokenType specifies the type of the subject token being exchanged.
-	// Common values: oauth.TokenTypeAccessToken (default), oauth.TokenTypeIDToken, oauth.TokenTypeJWT.
-	// If empty, defaults to oauth.TokenTypeAccessToken.
+	// Common values: oauthproto.TokenTypeAccessToken (default), oauthproto.TokenTypeIDToken, oauthproto.TokenTypeJWT.
+	// If empty, defaults to oauthproto.TokenTypeAccessToken.
 	SubjectTokenType string `json:"subject_token_type,omitempty"`
 
 	// HeaderStrategy determines how to inject the token
@@ -195,8 +195,8 @@ func CreateMiddlewareFromHeader(config Config) (types.MiddlewareFunction, error)
 // This is the recommended approach for external authentication flows (OAuth/OIDC).
 //
 // The middleware will automatically select the appropriate token based on config.SubjectTokenType:
-//   - oauth.TokenTypeAccessToken: Uses token.AccessToken
-//   - oauth.TokenTypeIDToken or oauth.TokenTypeJWT: Uses token.Extra("id_token")
+//   - oauthproto.TokenTypeAccessToken: Uses token.AccessToken
+//   - oauthproto.TokenTypeIDToken or oauthproto.TokenTypeJWT: Uses token.Extra("id_token")
 //
 // This moves the token selection logic into the middleware where it belongs,
 // keeping the command layer focused on configuration.
@@ -210,11 +210,11 @@ func CreateMiddlewareFromTokenSource(
 
 	// Validate SubjectTokenType early to catch configuration errors at startup
 	if config.SubjectTokenType != "" &&
-		config.SubjectTokenType != oauth.TokenTypeAccessToken &&
-		config.SubjectTokenType != oauth.TokenTypeIDToken &&
-		config.SubjectTokenType != oauth.TokenTypeJWT {
+		config.SubjectTokenType != oauthproto.TokenTypeAccessToken &&
+		config.SubjectTokenType != oauthproto.TokenTypeIDToken &&
+		config.SubjectTokenType != oauthproto.TokenTypeJWT {
 		return nil, fmt.Errorf("invalid SubjectTokenType: %s (must be one of: %s, %s, %s)",
-			config.SubjectTokenType, oauth.TokenTypeAccessToken, oauth.TokenTypeIDToken, oauth.TokenTypeJWT)
+			config.SubjectTokenType, oauthproto.TokenTypeAccessToken, oauthproto.TokenTypeIDToken, oauthproto.TokenTypeJWT)
 	}
 
 	// Create a SubjectTokenProvider that handles token selection based on config
@@ -226,7 +226,7 @@ func CreateMiddlewareFromTokenSource(
 
 		// Select appropriate token based on configured type
 		switch config.SubjectTokenType {
-		case oauth.TokenTypeIDToken:
+		case oauthproto.TokenTypeIDToken:
 			// Extract ID token from Extra field (standard OIDC approach)
 			idToken, ok := token.Extra("id_token").(string)
 			if !ok || idToken == "" {
@@ -235,7 +235,7 @@ func CreateMiddlewareFromTokenSource(
 			}
 			return idToken, nil
 
-		case "", oauth.TokenTypeAccessToken:
+		case "", oauthproto.TokenTypeAccessToken:
 			// Use access token (default)
 			if token.AccessToken == "" {
 				slog.Error("Access token not available")
