--- pkg/oauth/doc.go
@@ -1,18 +1,0 @@
-// Copyright 2025 Stacklok, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// Package oauth provides shared RFC-defined types, constants, and validation utilities
-// for OAuth 2.0 and OpenID Connect. It serves as a shared foundation for both OAuth
-// clients and servers, including redirect URI validation per RFC 6749 and RFC 8252.
-package oauth
