+++ pkg/oauthproto/redirect_test.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright 2025 Stacklok, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package oauth
+package oauthproto
 
 import (
 	"strings"
